Add NextAuth with Google

+++ app/actions/events.ts
@@ -3,6 +3,8 @@
 import {supabaseServer} from "@/supabase/supabaseServer";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
 
 export async function createEvent(datas: FormData) {
     const name = datas.get("name");
@@ -15,6 +17,9 @@ export async function createEvent(datas: FormData) {
 }
 
 export async function updateEvent(datas: FormData, id: string) {
+    const session = await getServerSession(authOptions);
+    if(!session) throw new Error("User is not logged");
+
     const name = datas.get("name");
     const desc = datas.get("desc");
     const start = datas.get("date_start");
@@ -25,6 +30,9 @@ export async function updateEvent(datas: FormData, id: string) {
 }
 
 export async function deleteEvent(id: number) {
+    const session = await getServerSession(authOptions);
+    if(!session) throw new Error("User is not logged");
+
     await supabaseServer.from("events").delete().eq("id", id);
     revalidatePath("/evenementiel");
 }
